requirements.txt
Django==5.2.1
python-decouple==3.8
langchain==0.3.25
langchain-core==0.3.60
langchain-groq==0.3.2
langchain-huggingface==0.2.0
langchain-chroma==0.2.4
langchain-community==0.3.24
langchain-text-splitters==0.3.8
chromadb==1.0.9
transformers==4.52.2
sentence-transformers==4.1.0
huggingface-hub==0.31.4
torch==2.7.0
pypdf==5.5.0
PyPDF2==3.0.1
